+++ src/_thrash/csp_example.js
