+++ src/_thrash/install.js
