+++ src/_thrash/solution.js
