+++ src/lib/useful.js
